refactor(tui): move choose-storage-file dialog tests inline; narrow API
Relocates the three startup-dialog tests from the integration test file
into #[cfg(test)] mod tests so they have crate-private access, and reverts
the pub use re-exports of AppMode and DialogMode that were added solely to
let those tests assert on app.mode from outside the crate.
